djangomongo/requirements.txt
dataclasses==0.6
Django==2.2.6
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
djongo==1.2.36
PyJWT==1.7.1
pymongo==3.9.0
pytz==2019.3
sqlparse==0.2.4
